requirements.txt
django-fields==0.3.0
pyOpenSSL==0.13.1
pycrypto==2.6
wsgiref==0.1.2
